WIP: implement postgres to add function
This is just a WIP commit to save my work. Code from the Commandhandler
needs to get pulled into services. Also it is very important to create
service methods which will take care of saving a user in the correct
order. Or I should think about activating cascading save and just take
care that all fields are set for the specific object which you want to
save/update.